backend/agents/environment-monitor/pyproject.toml
[project]
name = "environment_monitor_agent"
version = "0.1.0"
description = "An agent that monitors weather and air quality using Coral Protocol and LLM reasoning."
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.1",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-experimental==0.3.4",
    "langchain-groq==0.3.4",
    "langchain-mcp-adapters==0.1.7",
    "langchain-google-vertexai",
    "uv>=0.7.17",
]
